requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements/dev.in
#
appdirs==1.4.4
    # via black
asgiref==3.3.4
    # via django
backcall==0.2.0
    # via ipython
git+git://github.com/psf/black
    # via -r requirements/dev.in
click==7.1.2
    # via
    #   black
    #   pip-tools
decorator==5.0.7
    # via ipython
django-cors-headers==3.7.0
    # via -r requirements/base.in
django-debug-toolbar==3.2.1
    # via -r requirements/dev.in
django-environ==0.4.5
    # via -r requirements/base.in
django-extensions==3.1.3
    # via -r requirements/dev.in
django==3.2.2
    # via
    #   -r requirements/base.in
    #   django-cors-headers
    #   django-debug-toolbar
    #   django-extensions
    #   djangorestframework
djangorestframework==3.12.4
    # via -r requirements/base.in
ipython-genutils==0.2.0
    # via traitlets
ipython==7.23.1
    # via -r requirements/dev.in
jedi==0.18.0
    # via ipython
markdown==3.3.4
    # via -r requirements/base.in
matplotlib-inline==0.1.2
    # via ipython
mypy-extensions==0.4.3
    # via black
parso==0.8.2
    # via jedi
pathspec==0.8.1
    # via black
pep517==0.10.0
    # via pip-tools
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pip-tools==6.1.0
    # via -r requirements/dev.in
prompt-toolkit==3.0.18
    # via ipython
psycopg2==2.8.6
    # via -r requirements/base.in
ptyprocess==0.7.0
    # via pexpect
pygments==2.9.0
    # via ipython
pytz==2021.1
    # via django
regex==2021.4.4
    # via black
sqlparse==0.4.1
    # via
    #   django
    #   django-debug-toolbar
toml==0.10.2
    # via
    #   black
    #   pep517
traitlets==5.0.5
    # via
    #   ipython
    #   matplotlib-inline
wcwidth==0.2.5
    # via prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
